fix int

+++ project/module/models/load_model.py
@@ -49,7 +49,7 @@ def load_model(model_name, hparams=None):
     elif model_name == "mlp":
         net = SimpleMLP(dims, hparams.mlp_dim, hparams.target_dim)
     elif model_name == "linear":
-        net = LinearLayer(dims, hparams.target_dim, t, t/hparams.patch_size[-1])
+        net = LinearLayer(dims, hparams.target_dim, t, int(t/hparams.patch_size[-1]))
     elif model_name == "lstm":
         net = LSTM(dims, hparams.lstm_dim, hparams.lstm_layers, hparams.target_dim)
     else:
